requirements.txt
asgiref==3.2.3
Django==2.2.8
django-filter==2.2.0
django-google-maps==0.11.0
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
location==0.0.7
psycopg2-binary==2.8.4
pycparser==2.19
python-google-places==1.4.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
